mysite/requirements.txt
Django
wagtail
mysqlclient
django-environ
django-cors-headers
uwsgi
djangorestframework
